springmvc-sec/src/test/java/com/shsz/young/jeff/test/model/Car.java
package com.shsz.young.jeff.test.model;
import javax.validation.Valid;
import javax.validation.constraints.Min;
import javax.validation.constraints.NotNull;
import javax.validation.constraints.Size;

import com.shsz.young.jeff.test.validator.CaseMode;
import com.shsz.young.jeff.test.validator.CheckCase;

@SuppressWarnings("unused")
public class Car {

	//@NotNull
    //@Valid
    private Person driver;
	
	@NotNull
    private String manufacturer;

    @NotNull
    @Size(min = 2, max = 14)
    @CheckCase(CaseMode.UPPER)
    private String licensePlate;

    @Min(2)
    private int seatCount;
    
    public Car(String manufacturer, String licencePlate, int seatCount) {
        this.manufacturer = manufacturer;
        this.licensePlate = licencePlate;
        this.seatCount = seatCount;
    }

	public Person getDriver() {
		return driver;
	}

	public void setDriver(Person driver) {
		this.driver = driver;
	}



}